# ByteMarket/backend/src/main/java/es/grupo04/backend/api-docs/api-docs.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: https://localhost:8443
  description: Generated server url
paths:
...
  /api/v1/chats/{productId}:
    post:
      summary: Create a new chat for a product by its ID
      tags:
      - chat-rest-controller
      operationId: createChat
      parameters:
      - name: productId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ChatDTO"
  /api/v1/chats/{id}/send:
    post:
      summary: Send a message in a chat by its ID
      tags:
      - chat-rest-controller
      operationId: sendMessage
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: message
        in: query
        required: true
        schema:
          type: string
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/v1/chats/{id}/sell:
    post:
      summary: Mark a product as sold in a chat by its ID
      tags:
      - chat-rest-controller
      operationId: sellProduct
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
  /api/v1/chats:
    get:
      summary: Retrieve a list of all chats for the authenticated user
      tags:
      - chat-rest-controller
      operationId: getUserChats
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatDTO"
  /api/v1/chats/{id}:
    get:
      summary: Fetch a specific chat by its ID
      tags:
      - chat-rest-controller
      operationId: getChat
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ChatDTO"
...
components:
  schemas:
...
    UserBasicDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        image:
          type: string
        hasImage:
          type: boolean
...
    ProductBasicDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        price:
          type: number
          format: float
        category:
          type: string
        sold:
          type: boolean
        publishDate:
          type: string
          format: date
...
    ChatDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        product:
          $ref: "#/components/schemas/ProductBasicDTO"
        userBuyer:
          $ref: "#/components/schemas/UserBasicDTO"
        userSeller:
          $ref: "#/components/schemas/UserBasicDTO"
        messages:
          type: array
          items:
            $ref: "#/components/schemas/MessageDTO"
        selling:
          type: boolean
    MessageDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        sender:
          $ref: "#/components/schemas/UserBasicDTO"
        message:
          type: string
        sentAt:
          type: string
...